backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0

# Database (SQLModel + Postgres)
sqlmodel>=0.0.22
psycopg2-binary>=2.9.10

# Security (PyJWT + Direct Bcrypt)
pyjwt>=2.10.1
bcrypt>=4.2.0
python-multipart>=0.0.20

# Utilities
python-dotenv>=1.0.1